RunBox/apps/api/requirements/base.txt
fastapi==0.109.0
uvicorn[standard]==0.25.0
sqlalchemy==2.0.25
alembic==1.13.1
psycopg2-binary==2.9.9
pydantic-settings==2.1.0
pydantic==2.5.3
python-dotenv==1.0.0
minio==7.2.0
boto3==1.34.18
httpx==0.26.0
passlib==1.7.4
python-jose[cryptography]==3.3.0
orjson==3.9.10
email-validator==2.1.0.post1
